+++ index_trace/__init__.py
@@ -2,4 +2,4 @@
 IndexTrace package initialization.
 """
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"